+++ client/django-formset.monolith.ts
@@ -5,10 +5,10 @@ import {StyleHelpers} from './django-formset/helpers';
 // remember to always reflect imports below here also in django-formset.ts
 import {DjangoSelectizeElement} from './django-formset/DjangoSelectize';
 import {SortableSelectElement} from './django-formset/SortableSelect';
-import {DualSelectorElement} from "./django-formset/DualSelector";
-import {RichTextAreaElement} from "./django-formset/RichtextArea";
-import {DjangoSlugElement} from "./django-formset/DjangoSlug";
-import {DatePickerElement, DateTimePickerElement} from "./django-formset/DateTimePicker";
+import {DualSelectorElement} from './django-formset/DualSelector';
+import {RichTextAreaElement} from './django-formset/RichtextArea';
+import {DjangoSlugElement} from './django-formset/DjangoSlug';
+import {DateFieldElement, DatePickerElement, DateTimeFieldElement, DateTimePickerElement} from './django-formset/DateTime';
 
 
 window.addEventListener('DOMContentLoaded', (event) => {
@@ -25,10 +25,30 @@ window.addEventListener('DOMContentLoaded', (event) => {
 	customElementNames.push('django-richtext');
 	window.customElements.define('django-slug', DjangoSlugElement, {extends: 'input'});
 	customElementNames.push('django-slug');
+	window.customElements.define('django-datefield', DateFieldElement, {extends: 'input'});
+	customElementNames.push('django-datefield');
+	window.customElements.define('django-datecalendar', DateFieldElement, {extends: 'input'});
+	customElementNames.push('django-datecalendar');
 	window.customElements.define('django-datepicker', DatePickerElement, {extends: 'input'});
 	customElementNames.push('django-datepicker');
+	window.customElements.define('django-datetimefield', DateTimeFieldElement, {extends: 'input'});
+	customElementNames.push('django-datetimefield');
+	window.customElements.define('django-datetimecalendar', DateFieldElement, {extends: 'input'});
+	customElementNames.push('django-datetimecalendar');
 	window.customElements.define('django-datetimepicker', DateTimePickerElement, {extends: 'input'});
 	customElementNames.push('django-datetimepicker');
+	window.customElements.define('django-daterangefield', DateTimePickerElement, {extends: 'input'});
+	customElementNames.push('django-daterangefield');
+	window.customElements.define('django-daterangecalendar', DateTimePickerElement, {extends: 'input'});
+	customElementNames.push('django-daterangecalendar');
+	window.customElements.define('django-daterangepicker', DateTimePickerElement, {extends: 'input'});
+	customElementNames.push('django-daterangepicker');
+	window.customElements.define('django-datetimerangefield', DateTimePickerElement, {extends: 'input'});
+	customElementNames.push('django-datetimerangefield');
+	window.customElements.define('django-datetimerangecalendar', DateTimePickerElement, {extends: 'input'});
+	customElementNames.push('django-datetimerangecalendar');
+	window.customElements.define('django-datetimerangepicker', DateTimePickerElement, {extends: 'input'});
+	customElementNames.push('django-datetimerangepicker');
 
 	const foundIds = new Set<string>();
 	document.querySelectorAll('django-formset [id]').forEach(element => {