+++ formset/renderers/foundation.py
@@ -6,6 +6,7 @@ class FormRenderer(DefaultFormRenderer):
     framework = 'foundation'
 
     _template_mapping = dict(DefaultFormRenderer._template_mapping, **{
+        'django/forms/div.html': 'formset/foundation/form.html',
         'django/forms/default.html': 'formset/foundation/form.html',
         'django/forms/widgets/checkbox.html': 'formset/foundation/widgets/checkbox.html',
         'django/forms/widgets/radio.html': 'formset/foundation/widgets/multiple_input.html',