+++ formset/renderers/uikit.py
@@ -11,6 +11,7 @@ class FormRenderer(DefaultFormRenderer):
         super().__init__(**kwargs)
 
     _template_mapping = dict(DefaultFormRenderer._template_mapping, **{
+        'django/forms/div.html': 'formset/uikit/form.html',
         'django/forms/default.html': 'formset/uikit/form.html',
         'django/forms/widgets/checkbox.html': 'formset/uikit/widgets/checkbox.html',
         'django/forms/widgets/radio.html': 'formset/uikit/widgets/multiple_input.html',